Health care/app/requirements.txt
python-dotenv
streamlit
pandas
numpy
numpy==1.26.4
matplotlib
plotly
seaborn
scikit-learn
xgboost
lightgbm
catboost
tensorflow
torch
keras

datetime 
pillow
opencv-python 
keras


ollama
langchain
langchain-core
langchain-community
transformers 


langchain_openai
langchain_community





